Combat Message Overlay code cleanup
Cleaned up the new code and did some very light refactoring.
