+++ Magitek/Converters/BooleanToOpaqueConverter.cs
@@ -2,8 +2,6 @@
 {
     public sealed class BooleanToOpaqueConverter : BooleanConverter<double>
     {
-        public BooleanToOpaqueConverter() :
-            base(1.0, 0.0)
-        { }
+        public BooleanToOpaqueConverter() : base(1.0, 0.0) { }
     }
 }
